Now I check if the MGF files exist

+++ bash/systemhc/script.sh
@@ -15,15 +15,26 @@ for i in $(cat sample_list.txt); do
     if [ $(cat $csLoc/$i.txt | wc -l ) -eq 0 ]; then
        echo "ignoring $csLoc/$i.txt"
     else
-	cat $csLoc/$i.txt |  sed -e "s/^/$mgfLoc\//" | xargs awk 1 > $csLoc/$i.mgf
-	wget "https://systemhcatlas.org/mnt/Systemhc/Data/180409_build/$i/Spectrast/cons.tar.gz" -O "$csLoc/$i-cons.tar.gz" --no-check-certificate
-	mkdir $csLoc/$i-cons
-	tar xvf $csLoc/$i-cons.tar.gz -C $csLoc/$i-cons
-	tsvFile=$(find $csLoc/$i-cons -name "consensus.tsvh" | head -n 1)
-	cp $tsvFile $tsvhLoc/$i.tsvh
-	mgfFiles+=("$csLoc/$i.mgf")
-	tsvhFiles+=("$tsvhLoc/$i.tsvh")
-	length=$length+1
+	allMGFFilesExist=true
+	mgfsInSample=$( cat $csLoc/$i.txt |  sed -e "s/^/$mgfLoc\//" )
+	for x in $mgfsInSample; do
+	    if [ ! -f $x ]; then
+		allMGFFilesExist=false
+	    fi
+	done
+	if $allMGFFilesExist; then
+	    awk 1 $mgfsInSample > $csLoc/$i.mgf
+	    wget "https://systemhcatlas.org/mnt/Systemhc/Data/180409_build/$i/Spectrast/cons.tar.gz" -O "$csLoc/$i-cons.tar.gz" --no-check-certificate
+	    mkdir $csLoc/$i-cons
+	    tar xvf $csLoc/$i-cons.tar.gz -C $csLoc/$i-cons
+	    tsvFile=$(find $csLoc/$i-cons -name "consensus.tsvh" | head -n 1)
+	    cp $tsvFile $tsvhLoc/$i.tsvh
+	    mgfFiles+=("$csLoc/$i.mgf")
+	    tsvhFiles+=("$tsvhLoc/$i.tsvh")
+	    length=$length+1
+	else
+	    echo "JORDAN: Not all MGF files in $csLoc/$i.txt exist"
+	fi
     fi
 done
